change js in filters and csp.yaml #2609
