+++ config/app/csp.yaml
@@ -1,0 +1,18 @@
+parameters:
+  csp_parameters:
+    default-src: "'none'"
+    script-src: "'self' 'unsafe-inline' 'unsafe-eval' https://cdn.matomo.cloud/histologe.matomo.cloud/matomo.js https://cdn.jsdelivr.net/npm/leaflet@1.7.1/dist/leaflet.js  https://cdn.jsdelivr.net/npm/leaflet.markercluster@1.5.3/dist/leaflet.markercluster.js https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js https://cdn.jsdelivr.net/npm/tippy.js@6/dist/tippy-bundle.umd.js"
+    style-src: "'self' 'unsafe-inline' https://cdn.jsdelivr.net/npm/leaflet@1.7.1/dist/leaflet.css https://cdn.jsdelivr.net/npm/leaflet.markercluster@1.5.3/dist/MarkerCluster.Default.css https://cdn.jsdelivr.net/npm/leaflet.markercluster@1.5.3/dist/MarkerCluster.css https://cdn.jsdelivr.net/npm/tippy.js@6/themes/light-border.css"
+    style-src-attr: "'self' 'unsafe-inline'"
+    img-src: "'self' data: blob: https://voxusagers.numerique.gouv.fr https://*.tile.openstreetmap.org https://cdn.jsdelivr.net https://jedonnemonavis.numerique.gouv.fr"
+    connect-src: "'self' https://api-adresse.data.gouv.fr https://cdn.matomo.cloud https://histologe.matomo.cloud https://koumoul.com https://sentry.incubateur.net"
+    font-src: "'self'"
+    frame-src: "'none'"
+    object-src: "'none'"
+    base-uri: "'self'"
+    form-action: "'self'"
+    frame-ancestors: "'none'"
+    media-src: "'self'"
+    require-trusted-types-for: "'script'"
+
+
